# Swap randsize & dist conditionals